modify add area and shelter logic

+++ DTO/Animal/GetAnimal.cs
@@ -22,7 +22,7 @@
 
         public string Status { get; set; }
 
-        public short? AreaId { get; set; }
+        public int? AreaId { get; set; }
 
         public string Area { get; set; }
