+++ DTO/News/GetNews.cs
@@ -8,7 +8,7 @@
 
         public string Message { get; set; }
 
-        public short? AreaId { get; set; }
+        public int? AreaId { get; set; }
 
         public string Area { get; set; }
